Relocate static page layout
